Selectively load and register CoffeeScript.
Meaning, don't register .coffee and ._coffee handlers unless CoffeeScript's
actually being used -- that breaks the default contract of Node.

Now, when registering only, ._coffee is only registered if .coffee is, and
when running a file too, both are registered if run via _coffee.

This also fixes a regression I had introduced: when _coffee is used to run a
file that's .coffee and not ._coffee.