Load CoffeeScript relative to caller's directory!
This fixes Sage/streamlinejs#84 -- if Streamline is installed in a parent
directory, while CoffeeScript is installed in the current one, now Streamline
can correctly find CoffeeScript. Sweet!

Added a TODO to ideally apply a similar technique when compiling Coffee code
as part of require() -- it should use the instance of CoffeeScript that's
already been required, relative to the calling *module*, not the cwd.