Rename shebang CoffeeScript example extension.
@bjouhier: this is tricky -- underscored.js relies on the file extension to
determine not only whether this is Coffee or JS (this was the case before any
of my changes as well), but also whether this is Streamlined Coffee/JS or not.

This means that shebang scripts shouldn't have a `.sh` extension like the
readme currently suggests. That might be good anyway -- it's not actually a
shell script.

But it probably should be possible for shebang files to have *no* extension.
In that case, we could/should assume Streamlined Coffee/JS based on which
executable was run: _node or _coffee. I can make this change.

All sound reasonable to you?